Integrate protocol management into search overlay
