+++ src/DocFinder.App/Views/Controls/ProtocolListControl.xaml
@@ -1,0 +1,27 @@
+<UserControl x:Class="DocFinder.App.Views.Controls.ProtocolListControl"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:ui="http://schemas.lepo.co/wpfui/2022/xaml">
+    <DockPanel>
+        <StackPanel Orientation="Horizontal" DockPanel.Dock="Top">
+            <ui:TextBox x:Name="TitleFilter" Width="120" Margin="2" PlaceholderText="Title" TextChanged="OnFilterChanged"/>
+            <ui:TextBox x:Name="RefFilter" Width="100" Margin="2" PlaceholderText="Reference" TextChanged="OnFilterChanged"/>
+            <ui:TextBox x:Name="PersonFilter" Width="120" Margin="2" PlaceholderText="Responsible" TextChanged="OnFilterChanged"/>
+        </StackPanel>
+        <ui:DataGrid x:Name="protocolsGrid" AutoGenerateColumns="False" CanUserAddRows="False" CanUserDeleteRows="False" IsReadOnly="False"
+                     RowEditEnding="ProtocolsGrid_RowEditEnding">
+            <ui:DataGrid.Columns>
+                <DataGridTextColumn Header="Title" Binding="{Binding Title}"/>
+                <DataGridTextColumn Header="Reference" Binding="{Binding ReferenceNumber}"/>
+                <DataGridTextColumn Header="Issued" Binding="{Binding IssuedBy}"/>
+                <DataGridTextColumn Header="Responsible" Binding="{Binding ResponsiblePerson}"/>
+                <DataGridTextColumn Header="Date" Binding="{Binding IssueDateUtc}"/>
+            </ui:DataGrid.Columns>
+            <ui:DataGrid.ContextMenu>
+                <ContextMenu>
+                    <MenuItem Header="Open File" Click="OpenFile_Click"/>
+                </ContextMenu>
+            </ui:DataGrid.ContextMenu>
+        </ui:DataGrid>
+    </DockPanel>
+</UserControl>
